Add files via upload
Go to IMDb.com, unfold the Menu button and navigate to the Born today section;
delete default search, then unfold Birth date and search for Celebrities born the same
day as today but exactly 40 years ago, using the date picker for the “from” option and
the string field for the “to” option. Now on the 1st result in the list, click on the 1st link
you can find on the description (if any) and take a screenshot.